requirements.txt
Django==3.0.7
django-cors-headers==3.4.0
django-storages==1.9.1
djangorestframework==3.11.0
SQLAlchemy==1.3.17
pandas==1.0.4
mysqlclient==1.4.6